refactor(frontend): extract sub-component from IcSendModal
